TW-2301: extract activity feed controller and metadata harvesting hook
