+++ src/activity/feed/index.ts
@@ -1,3 +1,5 @@
+export { createActivityFeedController } from './activity-feed-controller';
+export type { ActivityFeedController, ActivityFeedSourceFailure, ActivityFeedState } from './activity-feed-controller';
 export { createTezosActivitySource } from './tezos-source';
 export { createEvmActivitySource } from './evm-source';
 export { cutAtBoundary, getSafeBoundary, mergeActivityBuffers } from './merge';
